% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_area.R
\name{square_area}
\alias{square_area}
\title{Square area}
\usage{
square_area(length, width)
}
\arguments{
\item{length}{The length of the square [m]}

\item{width}{The width of the square [m]}
}
\description{
Function to calculate the area of a square. Requires numeric input.
}
\examples{
square_area(4,5)

\dontrun{
square_area(4,"5")}
}
\author{
Marieke Dirksen
}
